requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
APScheduler==3.11.0
attrs==24.3.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2024.12.14
charset-normalizer==3.4.1
discord==2.3.2
discord.py==2.4.0
et_xmlfile==2.0.0
frozenlist==1.5.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
httplib2==0.22.0
idna==3.10
multidict==6.1.0
numpy==2.3.1
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.1
propcache==0.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
yarl==1.18.3
